+++ changelog.md
@@ -1,2 +1,11 @@
-- 2020-08-20 - Bugfixes and split view
-- 2020-05-11 - Initial beta release
+- 2020-09-03: Improved handling of enabled/disabled komments
+
+  - kommentsAreEnabled() page methods
+  - disable komments after certain amount of time
+
+- 2020-08-20:
+
+  - Komments split view
+  - bugfixes
+
+- 2020-05-11: Initial beta release